CC08, and cleaned up folder structure
