+++ README.md
@@ -20,7 +20,4 @@
 
 ### 401
 
-1. [Challenge - Insert and Shift Array](challenges/ArrayShift)
-1. [Challenge - Linked List](challenges/Linked-Lists/Linked-Lists)
-1. [Challenge - Linked List Insertions](challenges/LinkedListInsertions)
-1. [Challenge - Kth From End](challenges/LinkedListKthFromEndTest)
+1. [Challenge - LinkedLists](challenges/LinkedLists)
