--- challenges/Libraries/LLLibrary/LLLibrary.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30309.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LLLibrary", "LLLibrary.csproj", "{FCAA3936-6E00-47EA-BE99-9B91893262D7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FCAA3936-6E00-47EA-BE99-9B91893262D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FCAA3936-6E00-47EA-BE99-9B91893262D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FCAA3936-6E00-47EA-BE99-9B91893262D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FCAA3936-6E00-47EA-BE99-9B91893262D7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {206FBA1F-C744-46AE-B094-5B902AB2C699}
-	EndGlobalSection
-EndGlobal
