--- challenges/Linked-Lists/Linked-Lists/Linked-Lists.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30225.117
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedList", "LinkedList.csproj", "{27404473-4081-4DF5-B8C9-2A5F957CBCD5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListTest", "..\LinkedListTest\LinkedListTest.csproj", "{8990989E-AA16-47B8-A0B7-7D16411C8DAD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{27404473-4081-4DF5-B8C9-2A5F957CBCD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{27404473-4081-4DF5-B8C9-2A5F957CBCD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{27404473-4081-4DF5-B8C9-2A5F957CBCD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{27404473-4081-4DF5-B8C9-2A5F957CBCD5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8990989E-AA16-47B8-A0B7-7D16411C8DAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8990989E-AA16-47B8-A0B7-7D16411C8DAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8990989E-AA16-47B8-A0B7-7D16411C8DAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8990989E-AA16-47B8-A0B7-7D16411C8DAD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D61B648A-88C4-4447-91C6-EFAE72B13C30}
-	EndGlobalSection
-EndGlobal
