--- challenges/Linked-Lists/LinkedListTest/LinkedListTest.cs
@@ -1,106 +1,0 @@
-using System;
-using Xunit;
-using LLLibrary;
-
-namespace LinkedListTest
-{
-    public class LinkedListTest
-    {
-        [Fact]
-        public void CanInstantiateEmptyLinkedList()
-        {
-            LinkedList list = new LinkedList();
-            Assert.Null(list.Head);
-        }
-
-        [Fact]
-        public void CanInsertIntoLinkedList()
-        {
-            // insert means add to the front
-            //Arrange
-            LinkedList list = new LinkedList();
-            list.Insert(4);
-            Assert.Equal(4, list.Head.Value);
-        }
-
-        [Fact]
-        public void CanInsertMultipleNodesInLinkedList()
-        {
-            LinkedList list = new LinkedList();
-            list.Insert(4);
-            list.Insert(8);
-            list.Insert(15);
-            list.Insert(16);
-            list.Insert(23);
-            list.Insert(42);
-
-            Assert.Equal(42, list.Head.Value);
-
-        }
-
-        [Fact]
-        public void CanFindValueInList()
-        {
-            // arrange
-            LinkedList list = new LinkedList();
-
-            list.Insert(4);
-            list.Insert(8);
-            list.Insert(15);
-            list.Insert(16);
-            list.Insert(23);
-            list.Insert(42);
-
-            // Act 
-            int searchForValue = 15;
-            bool exists = list.Includes(searchForValue);
-
-            // Assert
-            Assert.True(exists);
-
-        }
-
-        [Fact]
-        public void CannotFindValueInLinkedList()
-        {
-            // arrange
-            LinkedList list = new LinkedList();
-
-            list.Insert(4);
-            list.Insert(8);
-            list.Insert(15);
-            list.Insert(16);
-            list.Insert(23);
-            list.Insert(42);
-
-            // Act 
-            int searchForValue = 100;
-            bool exists = list.Includes(searchForValue);
-
-            // Assert
-            Assert.False(exists);
-
-        }
-
-        [Fact]
-        public void CanReturnAllValuesInLinkedList()
-        {
-            // Arrange
-            LinkedList list = new LinkedList();
-            list.Insert(4);
-            list.Insert(8);
-            list.Insert(15);
-            list.Insert(16);
-            list.Insert(23);
-            list.Insert(42);
-
-            // Act
-            string value = list.ToString();
-
-            string expected = "42 -> 23 -> 16 -> 15 -> 8 -> 4 -> NULL";
-            // Assert
-            Assert.Equal(expected, value);
-
-        }
-    }
-}
