--- challenges/LinkedListInsertions/LinkedListInsertionTest/LinkedListInsertion.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30225.117
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LinkedListInsertionTest", "LinkedListInsertionTest.csproj", "{704EF5F5-9621-4FF2-8E93-D5079BB43A2F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{704EF5F5-9621-4FF2-8E93-D5079BB43A2F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{704EF5F5-9621-4FF2-8E93-D5079BB43A2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{704EF5F5-9621-4FF2-8E93-D5079BB43A2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{704EF5F5-9621-4FF2-8E93-D5079BB43A2F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {31833F0F-6B0B-4A27-9FE7-F1FEE7599FBE}
-	EndGlobalSection
-EndGlobal
