--- challenges/LinkedListKthFromEndTest/LinkedListKthFromEndTest.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-    <PackageReference Include="coverlet.collector" Version="1.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="LLLibrary">
-      <HintPath>..\Libraries\LLLibrary\bin\Debug\netcoreapp3.1\LLLibrary.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-</Project>
