--- challenges/LinkedListKthFromEndTest/LinkedListKthFromEndTest.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30309.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinkedListKthFromEndTest", "LinkedListKthFromEndTest.csproj", "{B6336A30-2F09-47B6-980C-CBC175D1F167}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B6336A30-2F09-47B6-980C-CBC175D1F167}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B6336A30-2F09-47B6-980C-CBC175D1F167}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B6336A30-2F09-47B6-980C-CBC175D1F167}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B6336A30-2F09-47B6-980C-CBC175D1F167}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A6DB596C-FA8A-4986-BC76-8013B5DFB7C6}
-	EndGlobalSection
-EndGlobal
