+++ challenges/LinkedLists/Classes/Node.cs
