+++ challenges/LinkedLists/LinkedLists.csproj
@@ -3,13 +3,10 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Linked_Lists</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <Reference Include="LLLibrary">
-      <HintPath>..\..\Libraries\LLLibrary\bin\Debug\netcoreapp3.1\LLLibrary.dll</HintPath>
-    </Reference>
+    <Folder Include="Classes\" />
   </ItemGroup>
 
 </Project>
