+++ challenges/LinkedLists/Program.cs
